V7.12.4 latest code all update

+++ app/src/main/java/com/contusfly/utils/PickFileUtils.kt
@@ -123,7 +123,7 @@ object PickFileUtils {
     fun checkFileSize(filePath: String?, maxSize: Int): Boolean {
         if (filePath.isNullOrEmpty()) return true
         val file = File(filePath)
-        val fileInKb = file.length().toInt() / 1024
+        val fileInKb = file.length() / 1024
         val fileInMb = fileInKb / 1024
         return maxSize > fileInMb
     }
